requirements.txt
python-socketio>=5.8.0
eventlet>=0.33.3 